# changes call from readline() to read1() to speedup output syncing