## Changes proposed in this PR
- changes call from readline() to read1() to speedup output syncing

## Why are we making these changes?

The stdout and stderr syncer is much slower than the Mesos `CommandExecutor` version. We wish to lower the performance gap by optimizing the stdout/stderr writers.

## Additional notes

- the command python3 -c '[print("Hello:",i) for i in range(10000000)]' runtime is down to 240 secs from 840 secs
 - the mesos command executor runs in 120 secs

Command:
python3 -c '[print("Hello:",i) for i in range(100000000)]'

=== Job: fe785ef0-f8c1-41a3-adac-deb32c7fb487 (shamsimam_job) ===
Executor     **mesos**                                    
Job Instance                          Run Time                                          Host         Instance Status
4e7f890f-3ccc-423c-9cb3-1875d184144f  **1 minute and 54 seconds** (started 26 minutes ago)  localhost-3  Success

=== Job: a11173f1-d302-444c-a26c-d604bb67cfb1 (shamsimam_job) ===
Executor     **cook (with readline version)**
Job Instance                          Run Time                                            Host         Instance Status
671aaafc-b54a-4d04-91f8-468fa48591b3  **13 minutes and 57 seconds** (started 15 minutes ago)  localhost-3  Success

=== Job: fae0f771-b6e9-4d8b-b479-500a05223479 (shamsimam_job) ===
Executor     **cook (with read1 version)**
Job Instance                          Run Time                                           Host         Instance Status
40c8e2f4-d0d7-4736-a2bc-d4f7fb2d2f27  **3 minutes and 58 seconds** (started 22 minutes ago)  localhost-3  Success