@DaoWen 2017-12-07
I think this is good to merge as-is, but if we're still concerned about the overheads in the future, we can revisit the idea of using [`readinto1`](https://docs.python.org/3/library/io.html#io.BufferedIOBase.readinto1) with [memory views](https://docs.python.org/dev/library/stdtypes.html#memory-views) over the byte array (for writing just the portion of the byte-buffer that has new data).